Implemented flitering map with filter function
